+++ modules/halo/src/lua_compression.c
@@ -288,11 +288,6 @@ static int compression_service_event_handler(halo_lua_event_t event, void *user_
 		decompress_state.callback_ref = LUA_NOREF;
 		break;
 		
-	case HALO_LUA_EVENT_INTERRUPT:
-		/* Clear callback on Ctrl+C */
-		decompress_state.callback_ref = LUA_NOREF;
-		break;
-	
 	default:
 		break;
 	}